rename slack-commons.js to lib.js
