+++ src/appmixer/slack/list/CreateChannel/CreateChannel.js
@@ -1,5 +1,5 @@
 'use strict';
-const commons = require('../../slack-commons');
+const commons = require('../../lib');
 const { SlackAPIError } = require('../../errors');
 
 /**
