+++ src/appmixer/slack/list/ListPrivateChannels/ListPrivateChannels.js
@@ -1,5 +1,5 @@
 'use strict';
-const commons = require('../../slack-commons');
+const commons = require('../../lib');
 const { SlackAPIError } = require('../../errors');
 
 /**
