+++ src/appmixer/slack/reactions/ListReactions/ListReactions.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const { WebClient } = require('@slack/web-api');
-const commons = require('../../slack-commons');
+const commons = require('../../lib');
 
 const outputPortName = 'out';
